dom/os/test/test_remote.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Opening file from mozbrowser remote iframe</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1141021">os tests</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>

  <script>
  function runTest() {
    var iframe = document.createElement('iframe');
    iframe.src = 'remote.html';
    SpecialPowers.wrap(iframe).mozbrowser = true;
    iframe.setAttribute('remote', true);

    // abuse alert() as a way of doing IPC
    iframe.addEventListener('mozbrowsershowmodalprompt', function(e) {
      try {
        var data = JSON.parse(e.detail.message);
      }
      catch(ex) {
        dump('data was no proper JSON: ' + e.detail.message + '\n');
        ok(false, 'Response is JSON');
        return SimpleTest.finish();
      }

      is(data.hasFile, true, 'hasFile');

      var expected = 'This is data that you can read through the APIs!';
      var expectedData = expected.split('').map(c=>c.charCodeAt(0)).join(',');
      is(data.data, expectedData, 'data');

      SimpleTest.finish();
    }, false);

    document.body.appendChild(iframe);
  }

  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPermissions([
    { "type": "browser", "allow": 1, "context": document }
  ], () => {
    SpecialPowers.pushPrefEnv({"set": [
      [ "dom.mozBrowserFramesEnabled", true ],
      [ "dom.os.security.disabled", true ]
    ]}, runTest);
  });
  </script>
</body>
</html>
